Added super weapons
New icons for super weapons.
Added cases to recognise storm cannon and intel center.
New layer for super weapons
New tree just for super weapons
Super weapons on the structures pane.

This branch is based on Seth's master branch which includes shard 2 code that is not in his dev branch